Remove support for decimals
Decimal support (for IEEE 754-2008 32/64/128 decimals) is not
implemented in QPid and SwiftMQ (the classes are there, but broken or
not complete).

Encoding for decimals as required in the AMQP 1.0 is a non-trivial task
and most tasks in this context could be achieved with some workaround
(using a double/float/string representation for transfering the
information and then using BigDecimal for the operations).

So AMQP 1.0 decimals are not supported at this point.