Use committed offset on consumer recovery
The consumer coordinator tracks the offset of the last delivered message
and use it to recover for consumers without a name. This offset is only
kept in memory on the client side. With this commit, the coordinator
uses a committed offset (if any) in favor of the in-memory offset.

This can lead to restart of bit backward, but at least with the
guarantees the user expects with regards to offset tracking.