Add --delete-streams flag to performance tool
Streams not deleted by default at the end of a run. The
--delete-streams flag can be used to force-delete them.

The consumer default offset is now next. This is consistent with
the Consumer API and AMQP 0.9.1. Moreover when running on existing
streams both consumers and producers will start "working" at
the end of the streams (consumers will consume the messages
the producers just published).