Output warning if stream already exists with different properties
In performance tool.

The previous behavior was to crash, but it can be hard to have
the exact properties as on the server (created with the management UI
with explicit values, with a client, etc). Even a missing property
(e.g. leader-locator) could make the run crash.

Displaying a warning is good enough as most of the time the user
just wants to run the test against the existing stream and does
not want to delete it manually before the run.

They will now know that the stream does not have the same properties,
especially WRT to retention, which should to draw attention on
not running the test for too long with the appropriate retention.