Make RoutingStratey a public API
The metadata of the super stream are now provided in the
route method. This abstraction hides the infrastructure
needed to find out about the super stream topology.

The user can now have full control over the routing strategy,
even though the provided implementations ("hash" and "key")
should be enough most of the time.

The configuration of the routing strategy has been also refactored
to reflect the change and to be simpler.