Add SAC test suite, to launch from IDE
The SAC tests are scattered across different classes, because
SAC support must be tested at different layers of the client
(low-level client for framing and protocol notifications,
"simple" SAC consumer on a single stream, and as part of
a super stream). The test classes and methods are annotated
with a custom annotation and a test suite selects only those
classes and methods, to launch the SAC-related tests from
a single place in the IDE. The SAC test suite is excluded
in the plugin configuration though, because all the tests,
including the SAC ones, are run in the Maven build.