Avoid deadlock when subscribing/unsubscribing from multiple threads
Manager and manager pool in consumers coordinator used synchronized
block and called each other method which were themselves synchronized,
causing deadlock when the calls where interleaved.

This commit replaces the synchronized methods in manager by
a synchronized block on the manager pool (aka "owner"). This avoids
the deadlock at the cost of making managers waiting on their owner
lock. This should be fine in terms of performance as those classes
are called only when subscribing/unsubscribing, which are synchronous
operations. The cost of the network communication should always
be higher than the cost of synchronization.

Fixes #49