Reset subscription client on metadata update
A subscription is always cancelled on a metadata update (stream deleted or
Erlang process of the stream member restarted), so there is no guarantee
the consumer will keep the same client instance (= connection).
We're better off setting it to null, which set the SAC active flag to false
as well.
This avoids also an unnecessary warning when the consumer is re-activated
(otherwise it thinks it's already active).

Fixes #203