Retry offset retrieval in case connection to leader is down
All stream replicas are restarted (these are Erlang processed)
when the stream leader process restarts (crash or owning node stopped),
so the connection to the new leader may not be available when
a consumer needs to retrieve its offset (e.g. during the consumer
update for a SAC).

This commit takes this into account and retries the stored offset
retrieval in case the connection to the leader is not available yet.
The retry backoff policy is based on the one from the environment
for connection recovery.

Fixes #204