Improve recovery
Create a locator connection for each provided URI. This way
a connection can take over when the locator node goes down.
This speeds up recovery.

Track scheduled tasks. This is likely to be disabled in
a stable release. Useful to track down unfinished tasks.

Add retry to operations in the consumer coordinator.

Refresh consumer candidate nodes if the re-assignment of
a consumer times out.

This improvements are based on the feedback from the effects
of a rolling restart in K8S using stream-perf-test. Not all
producers and consumers are recovered after all nodes have
been restarted. The changes in this commit mitigates this problem.