Remove per-node data structure in producer coordinator
It would contain the client connections for a given
node. It was difficult to maintain the consistency
between those structures, so we're better off not using
this layer.

Now client connections are looked up with a linear scan,
which is good enough, as the number of connections
should remain under the thousand.