Use different executor factories for producers/consumers
Executor services are now shared between connection, but
there must be a distinction between producer and consumer
connections, as a consumer can depend on a RPC in
a producer connection (e.g. getting the stored offset
from a connection to the stream leader). So the producer
and consumer connections cannot share the same executors
to avoid deadlocks.

References #298