Add hints to error message for UnknownHostException
An UnknownHostException when creating a producer or
a consumer is usually due to a problem in the topology
discovery mechanism. A load balancer between the application
and the cluster is the most common reason, so this commit
makes the error message for UnknownHostException hints towards
this and provide links to appropriate resources.

This is a reasonably complex mechanism and we can't expect
all users to have been through all the documentation and blog posts
before stumbling on such a scenario, especially when considering
it happens on most K8S installations.