Add forceReplicaForConsumer flag
To force the use of a replica for consumers. The library
tries 5 times before falling back to the stream leader when
the flag is set to true (default is false, to favor faster
recovery).

The stream leader is started first and a replica can start
a few seconds later, so with bad timing a consumer can see
only the leader when it recovers. Setting the flag to true
will force the library to retry a few times to use a replica
if one becomes available.

The performance tool has also the corresponding flag now:
--force-replica-for-consumers.

This commit also adds the following options to the performance tool:
--hearbeat, --connection-recovery-interval,
--topology-recovery-interval.

Fixes #426, #427