Send empty message body instead of null message body
Null message body is forbidden in the AMQP 1.0 specification.
We set a data section with an empty binary when the message
body has not been set.

Stream messages are not checked on the server side, but
messages from other protocols will be checked for a non-null
body in RabbitMQ 4.0, and letting null body stream messages
through will crash when those messages are read with other protocols.

References rabbitmq/rabbitmq-server#6835
Fixes #544