Notify SAC when it is no longer active
A single active consumer may not always receive a notification from the
broker when it gets inactive. An obvious reason is the consumer
connection goes down. It is still possible to call the consumer update
listener from the library, which can help applications take an
appropriate action when a consumer goes from active to inactive.

This commit implements the call to the listener under such circumstances
(connection closed, stream unavailable because restarted, normal
consumer closing).