Log warning in case of unexpected use of dev mode
The dev mode is activated when no address resolver is set, the host is
localhost, and the user is guest. The dev mode sets a resolver that
always uses localhost, this avoids problems when e.g. the metadata frame
returns the hostname of a local container that the client application
cannot reach.

This commit adds a warn-level log message if the client detects a local
cluster setup (because there are several addresses passed in to the
address resolver).

References #420