Support OAuth2 authentication
This commit introduces OAuth2 settings (token endpoint URI, client
secret, etc) in the stream environment builder. When OAuth2 is in use,
the library fetches a token from the HTTP(S) endpoint and use it to
authenticate against the broker. The library also takes care of
refreshing tokens: when a token is about to expire, a new one is
fetched, and connections re-authenticate.

Fixes #781