Fix race condition on connection closing
The broker can close a connection abruptly (in case of re-authentication
fails) and the application can call Client#close() at the same time.
There can be a race condition whereby the call to close() triggers a
regular closing initiated by the client but the Netty channel is already
inactive. The call to close() then blocks until it reaches the RPC
timeout.

This commit makes sure the channel is still active before trying to send
the close command to the broker.